requirements.txt
Django
djangorestframework
mysqlclient
django-cors-headers
django-mysql
pymysql
python-dotenv
pytest
pytest-django
pymemcache